+++ lib/blocktrain/aggregations/car_weight_aggregation.rb
@@ -1,0 +1,31 @@
+module Blocktrain
+  module Aggregations
+    class CarWeightAggregation < Aggregation
+
+      def aggs
+        {
+          weight: {
+            terms: { field: 'memoryAddress' },
+            aggregations: {
+              avg_weight: {
+                avg: {
+                  field: 'value'
+                }
+              },
+              max_weight: {
+                max: {
+                  field: 'value'
+                }
+              },
+              min_weight: {
+                min: {
+                  field: 'value'
+                }
+              }
+            }
+          }
+        }
+      end
+    end
+  end
+end
